refactor: extract editor feature folder
Move src/editor/ into src/app/features/editor/{lib,widgets} and
absorb editor-adjacent UI: editor-toolbar, editor-stats-badge,
floating-actions, sidebar-edit-menu, sidebar-format-menu, the
find panel hook, editor-settings (renamed .ts), presence, and
the keyboard shortcuts setup from editor-utils.
