+++ src/app/features/backup/lib/scenarios.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from "vitest"
 import { co, Group } from "jazz-tools"
 import { createJazzTestAccount, setupJazzTestSync } from "jazz-tools/testing"
 import { UserAccount, Document, Asset, VideoAsset } from "@/schema"
-import { getPath } from "@/editor/frontmatter"
+import { getPath } from "@/app/features/editor/lib/frontmatter"
 import { getDocumentTitle } from "@/lib/document-utils"
 import type { BackupDoc } from "./sync"
 import { readManifest } from "./sync"
