+++ src/app/features/editor/lib/autocomplete-brackets.ts
