+++ src/app/features/editor/lib/editor-settings.ts
