+++ src/app/features/editor/lib/file-drop-cursor.ts
