+++ src/app/features/editor/lib/find-extension.ts
