+++ src/app/features/editor/lib/wikilink-decorations.ts
