+++ src/app/features/editor/widgets/editor.css
