+++ src/app/features/time-machine/screens/time-machine-screen.tsx
@@ -13,9 +13,9 @@ import {
 } from "lucide-react"
 import { toast } from "sonner"
 import { Document, UserAccount, Asset, ImageAsset, Space } from "@/schema"
-import { MarkdownEditor, useMarkdownEditorRef } from "@/editor/editor"
-import "@/editor/editor.css"
-import { useEditorSettings } from "@/lib/editor-settings"
+import { MarkdownEditor, useMarkdownEditorRef } from "@/app/features/editor"
+import "@/app/features/editor/widgets/editor.css"
+import { useEditorSettings } from "@/app/features/editor"
 import { getDocumentTitle } from "@/lib/document-utils"
 import { getSpaceGroup } from "@/app/features/spaces"
 import {
